+++ lib/govuk_publishing_components/presenters/machine_readable/faq_page_schema.rb
@@ -24,7 +24,9 @@ module GovukPublishingComponents
       end
 
       def questions_and_answers_markup
-        question_and_answers(page.body).map do |question, value|
+        question_and_answers(page.body)
+          .select { |_, value| value[:answer].present? }
+          .map do |question, value|
           q_and_a_url = section_url(value[:anchor])
 
           {
@@ -42,7 +44,7 @@ module GovukPublishingComponents
 
       # Generates a hash of questions and associated information:
       # - question: the text in the h2 tag preceding other markup. Questions are
-      #             used to key the hash. "Summary" is set as the default, as
+      #             used to key the hash. The page title is set as the default, as
       #             there is often a preamble in guides before any h2 is set.
       #
       # - :answer: the markup that is not an h2 tag. It is associated with the
@@ -53,7 +55,7 @@ module GovukPublishingComponents
       def question_and_answers(html)
         doc = Nokogiri::HTML(html)
 
-        question = "Summary"
+        question = page.title
 
         # rubocop:disable Style/IfInsideElse
         doc.xpath("html/body").children.each_with_object({}) do |element, q_and_as|